+++ e3nn_jax/_src/basic.py
@@ -9,6 +9,75 @@ from e3nn_jax._src.irreps import IntoIrreps
 from e3nn_jax._src.irreps_array import _infer_backend, _standardize_axis
 
 
+def from_chunks(
+    irreps: IntoIrreps,
+    chunks: List[Optional[jnp.ndarray]],
+    leading_shape: Tuple[int, ...],
+    dtype=None,
+    *,
+    backend=None,
+) -> e3nn.IrrepsArray:
+    r"""Create an IrrepsArray from a list of arrays.
+
+    Args:
+        irreps (Irreps): irreps
+        chunks (list of optional `jax.numpy.ndarray`): list of arrays
+        leading_shape (tuple of int): leading shape of the arrays (without the irreps)
+
+    Returns:
+        IrrepsArray
+    """
+    jnp = _infer_backend(chunks) if backend is None else backend
+
+    irreps = e3nn.Irreps(irreps)
+    if len(irreps) != len(chunks):
+        raise ValueError(
+            f"e3nn.from_chunks: len(irreps) != len(chunks), {len(irreps)} != {len(chunks)}"
+        )
+
+    if not all(x is None or isinstance(x, jnp.ndarray) for x in chunks):
+        raise ValueError(
+            f"e3nn.from_chunks: chunks contains non-array elements type={[type(x) for x in chunks]}"
+        )
+
+    if not all(
+        x is None or x.shape == leading_shape + (mul, ir.dim)
+        for x, (mul, ir) in zip(chunks, irreps)
+    ):
+        raise ValueError(
+            f"e3nn.from_chunks: chunks shapes {[None if x is None else x.shape for x in chunks]} "
+            f"incompatible with leading shape {leading_shape} and irreps {irreps}. "
+            f"Expecting {[leading_shape + (mul, ir.dim) for (mul, ir) in irreps]}."
+        )
+
+    for x in chunks:
+        if x is not None:
+            dtype = x.dtype
+            break
+
+    if dtype is None:
+        raise ValueError(
+            "e3nn.from_chunks: Need to specify dtype if chunks is empty or contains only None."
+        )
+
+    if irreps.dim > 0:
+        array = jnp.concatenate(
+            [
+                jnp.zeros(leading_shape + (mul_ir.dim,), dtype)
+                if x is None
+                else x.reshape(leading_shape + (mul_ir.dim,))
+                for mul_ir, x in zip(irreps, chunks)
+            ],
+            axis=-1,
+        )
+    else:
+        array = jnp.zeros(leading_shape + (0,), dtype)
+
+    zero_flags = tuple(x is None for x in chunks)
+
+    return e3nn.IrrepsArray(irreps, array, zero_flags=zero_flags)
+
+
 def as_irreps_array(array: Union[jnp.ndarray, e3nn.IrrepsArray], *, backend=None):
     """Convert an array to an IrrepsArray.
